radar-frontend/tools/requirements.txt
openpyxl>=3.1.2
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
unidecode>=1.3.8
